Add mtkrc path
Some init files instead of mentioning /vendor/etc/init/hw/ use ro.mtkrc.path. It is not a problem in stock as default.prop from vendor is used by init and it contains that prop. But in Custom ROMs default.prop in present in bootimage and is symlinked to /system/etc/prop.default. So adding this prop helps the rc files to locate other rc files, may even fix ril as ril rc files are called by init.mt6765.rc using this prop